Move events test module under cmd/integration_tests as well
